+++ src/PanicEngine/obj/Debug/net10.0/PanicEngine.AssemblyInfo.cs
@@ -14,7 +14,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyDescriptionAttribute("PanicEngine - A physics engine library")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("0.0.0.1")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("0.0.1+074b36bb5c6023146158919418fb1f6c8f22988c")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("0.0.1+1d667e65ee32fc21771be52af0829753036d0724")]
 [assembly: System.Reflection.AssemblyProductAttribute("PanicEngine")]
 [assembly: System.Reflection.AssemblyTitleAttribute("PanicEngine")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]