+++ tests/PanicEngine.Tests/obj/Debug/net10.0/PanicEngine.Tests.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("PanicEngine.Tests")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+074b36bb5c6023146158919418fb1f6c8f22988c")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+1d667e65ee32fc21771be52af0829753036d0724")]
 [assembly: System.Reflection.AssemblyProductAttribute("PanicEngine.Tests")]
 [assembly: System.Reflection.AssemblyTitleAttribute("PanicEngine.Tests")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]